Changed LICENSE to Creative Commons BY-SA 3.0
Rationale:

GPLv3 good for software, but not so much for documentation.  Several
specialized licenses for documentation exist. Among them, CC-BY-SA is
one of the most popular. In particular, it is used by Wikipedia, and
allows mixing in content to and from them.

For a human readable description of the license, see:

http://creativecommons.org/licenses/by-sa/3.0/

===

By commiting to the repo, you agree to license your work under the
terms of the LICENSE. Each person still retains copyright over their
own contributions, and is free to license them under any terms.

IANAL